requirements.txt
discord>=2.3.2
python-dotenv>=0.21.1
colorlog>=6.7.0
